gitlab: migrate existing nginx config to new location
155e8ca (gitlab/nginx: link config in new location, 2021-11-21) adjusted
the nginx configuration to a new location, but for existing setups, the
configuration still lives at the old location.

If the nginx config is missing, but still exists at the old location,
move it to the expected location.

+++ overlay/usr/local/bin/entrypoint.sh
@@ -43,8 +43,13 @@ install_conf() {
 			/etc/gitlab/logrotate/gitlab
 	fi
 	if [ ! -f "/etc/gitlab/nginx/http.d/default.conf" ]; then
-		install -Dm644 /home/git/gitlab/lib/support/nginx/gitlab \
-			/etc/gitlab/nginx/http.d/default.conf
+		if [ -f "/etc/gitlab/nginx/conf.d/default.conf" ]; then
+			# migrate from conf.d to http.d
+			mv /etc/gitlab/nginx/conf.d /etc/nginx/http.d
+		else
+			install -Dm644 /home/git/gitlab/lib/support/nginx/gitlab \
+				/etc/gitlab/nginx/http.d/default.conf
+		fi
 	fi
 }
